requirements.txt
cachetools
certifi
charset-normalizer
confluent-kafka
db-dtypes
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-core
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
httplib2
idna
numpy
packaging
pandas
proto-plus
protobuf
pyarrow
pyasn1
pyasn1-modules
pyparsing
python-dateutil
requests
rsa
six
uritemplate
urllib3
